+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # ChangeLog
 
 * Fixed Create Rotation Condition doesn't work in some cases.
+* Added Custom Instruction Keyframe handler for Geckolib Animations (Thanks to @Tcat20000)